requirements.txt
flask
pdfplumber
sentence-transformers
torch
scikit-learn
google-genai
python-dotenv
numpy